[fix] Student api 정리 및 수정

+++ src/main/java/com/TeachingManager/TeachingManager/Service/Student/StudentService.java
@@ -2,10 +2,12 @@ package com.TeachingManager.TeachingManager.Service.Student;
 
 import com.TeachingManager.TeachingManager.Repository.Student.StudentRepository;
 import com.TeachingManager.TeachingManager.Repository.User.Institute.InstituteRepository;
+import com.TeachingManager.TeachingManager.domain.CustomUser;
 import com.TeachingManager.TeachingManager.domain.Institute;
 import com.TeachingManager.TeachingManager.domain.Student;
 import com.TeachingManager.TeachingManager.DTO.Student.AddStudentRequest;
 import com.TeachingManager.TeachingManager.DTO.Student.UpdateStudentRequest;
+import com.TeachingManager.TeachingManager.domain.Teacher;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -32,28 +34,53 @@ public class StudentService {
     }
 
     //모든 학생 조회 메서드
-    public List<Student> findAll(){
-        return studentRepository.findAll();
+    public List<Student> findAll(CustomUser user){
+        if(user instanceof Teacher) {
+            throw new RuntimeException("선생님은 권한이 없습니다.");
+        }
+        else{
+            return studentRepository.findByInstitute_Pk(user.getPk());
+        }
     }
 
-    public Student findById(long id){
-        return studentRepository.findById(id)
+    
+    // 단일학생 조회 메서드
+    public Student findById(CustomUser user, long id){
+        Student student = studentRepository.findById(id)
                 .orElseThrow(()-> new IllegalArgumentException("not found: " + id));
+        if(student.getInstitute().getPk().equals(user.getPk())){
+            return student;
+        }
+        else{
+            throw new RuntimeException("올바르지 않은 접근입니다.");
+        }
     }
 
-    public void delete(long id){
-        studentRepository.deleteById(id);
+    // 학생 삭제 메서드
+    @Transactional
+    public String delete(CustomUser user, long id){
+        Student student = studentRepository.findById(id)
+                .orElseThrow(()-> new IllegalArgumentException("not found: " + id));
+        if(student.getInstitute().getPk().equals(user.getPk())){
+            studentRepository.deleteById(id);
+            return student.getName();
+        } // 학생의 소속학원에서 온 요청이 아닐 경우
+        else{
+            throw new RuntimeException("올바르지 않은 접근입니다.");
+        }
     }
 
     @Transactional
-    public Student update(long id, UpdateStudentRequest request) {
+    public Student update(CustomUser user, long id, UpdateStudentRequest request) {
         Student student = studentRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("not found: " + id));
 
-        student.update(request.getName(), request.getAge(), request.getGrade(), request.getPhoneNumber(), request.getParentName(), request.getParentNumber(), request.getGender(), request.getLevel());
-
-        return student;
+        if(student.getInstitute().getPk().equals(user.getPk())){
+            student.update(request.getName(), request.getAge(), request.getGrade(), request.getPhoneNumber(), request.getParentName(), request.getParentNumber(), request.getGender(), request.getLevel());
+            return student;
+        } // 학생의 소속
+       else{
+            throw new RuntimeException("올바르지 않은 접근입니다.");
+        }
     }
-
-
 }
